Add Qwaq enhancements, including methods required for HostWindowPlugin, from oscog branch.
Add realWindowHandle() macro to conform to image convention of treating window handle 1 as a reference to the main display window. Note: For X11, handles for HostWindowPlugin are currently *Window, but a handle registry could be added such that the handles are 1, 2, 3, ... as per original design intent. This would eliminate the need for special handling of handle 1.

Reformat additions to conform with README.Contributing guidelines.