Apply the changes of Mr Hachisuka for multibyte character composition. The diff is in the earlier email from Yoshiki:       http://lists.squeakfoundation.org/pipermail/squeak-dev/2016-May/189599.html       http://www2.asu.ac.jp/hachi/v3/scratch14ime.html
Changes are licenced MIT: http://lists.squeakfoundation.org/pipermail/squeak-dev/2016-May/189646.html

Adapt the patch to use the -compositioninput option of the X11 display module. Also fix invocation of that option, which was missing the necessary declaration in sqUnixMain.c (hence did not work).

With this change, if the -compositioninput options is on the command line, or if the equivalent SQUEAK_COMPOSITIONINPUT environment variable is set, then the recordPendingKeys() function in the X11 desplay module will use the enhancement of Mr. Hachisuka. Otherwise the original recordPendingKeys() logic is used.