VMMaker 4.15.11
Fix the interpreter for running images of the Squeak 3.6 and 3.8 era.
Several obsolete primitives were removed in VMMaker-dtl.387, but these
are still required for older images.

For this update, the primitive table initialiation is unchanged, but a
run time update is done for the ContextInterpreter to restore the old
primitive assignments. It is not clear if this will be a good strategy
but for now add InterpreterPrimitives>>installPrimitive:at: and use it
to update the table on first entry to the interpret() loop.

Possible future updates could support run time fixup to the primitive
table for Squeak 3.2 and earlier images.