VMMaker 4.16.2
Old image support: Rescue five sound primitive assignments for updatePrimitiveTable
for primitives that should be loaded from SouondGenerationPlugin, not from SoundPlugin.

The following ancient primitives are still not loadable into the primitive table
for old image support:

  #primWaveTableSoundmixSampleCountintostartingAtpan
  #primFMSoundmixSampleCountintostartingAtpan
  #primPluckedSoundmixSampleCountintostartingAtpan
  #primSampledSoundmixSampleCountintostartingAtpan
  #oldprimSampledSoundmixSampleCountintostartingAtleftVolrightVol

In addition #primitiveReadJoystick does not load from JoystickTabletPlugin
when running on Linux, but this is confirmed to be an artifact of the function
loader, which abandons the attempt to load a module if #initialiseModule answers
false, as is the case in the Unix stub code implementation (so it should work
on platforms that do support the joystick plugin).