Update ckformat.c per ImageFormat-dtl.39
Identify extended bytecodes for Sista in the image format number. Bit 10
identifies an image that contains extended bytecodes. Thus a 32 bit Spur
image is 6521, with Sista it is 7033, and a 64 bit Spur image is 68021,
with Sista it is 68533.

It is expected that additional bytecode sets can be identified by an
additional field in the image header, probably 32 bits containing two
16 bit fields, if both zero then Sista.

Per discussion with Eliot and Bert in a Squeak oversight board meeting
2019-05-15.