Add a cmake/configure option to specify name of the installed squeak script:
  --scriptname=<script> install squeak run script as <prefix>/bin/<script>

For example, if --scriptname=squeakvm is specified, then the run script
will be installed as /usr/local/bin/squeakvm rather than /usr/local/bin/squeak.
A corresponding update will be provided for the opensmalltalk-vm build
configuration in order to provide a mechanism for avoiding name collisions
when multiple types of VM are installed in the standard Unix directory structure.
Does not affect naming of squeak.sh, which may be subject to similar name
space conflicts.